feat: improve error handling for pubsub pkg
if pubsub subscribers fail or receive an error, the error would be
logged and the worker would exit without a signal that an error occurred
or an attempt to retry.  this adds both connection retry to the
subscribers and also sends an error to an error channel that can be
picked up by the caller.