fix: improve acquisition of lock
better handling of multiple go routines trying to acquire the lock and
trying to keep the number of mutexes in memory to a minimum. was causing
some issues if multiple goroutines were waiting for the lock and the
mutex was deleted from the map before the lock was reacquired.

this tracks the number of waiting goroutines and only cleans the mutex out of
the map when the number of waiting goroutine is zero.