Improved Lock Acquisition Handling (#59)
* fix: improve acquisition of lock

better handling of multiple go routines trying to acquire the lock and
trying to keep the number of mutexes in memory to a minimum. was causing
some issues if multiple goroutines were waiting for the lock and the
mutex was deleted from the map before the lock was reacquired.

this tracks the number of waiting goroutines and only cleans the mutex out of
the map when the number of waiting goroutine is zero.

* chore: Add redis to circleci deploy for testing

* fix: fix namespace; config 2x redis

* chore: disable dlc to save some credits

* fix: race condition in jobs pkg

all of sudden appeared in tests? weird.

* fix: lock around local map needed

* fix: lock around retries

in rare cases, a race condition occurs when the map is checked by two
processes simultaneously leading to a lock failure. this locks the
section where the map is checked and the action on the lock is performed
before unlocking.